la totale

+++ source/src/Core/ECS/Entity.cpp
@@ -13,7 +13,7 @@ void Entity::OnInspector(uint32_t entityID , Scene* scene)
 	Entity* ent = scene->GetEntities(entityID);
 
 	
-		for (size_t i = 0; i < ent->entityComponents.size(); i++)
+		for (uint32_t i = 0; i < ent->entityComponents.size(); i++)
 		{
 			if (ent->entityComponents[i] == ComponentNULL) continue;
 
@@ -22,7 +22,7 @@ void Entity::OnInspector(uint32_t entityID , Scene* scene)
 
 				std::vector<uint8_t>* data = scene->GetComponentDataArray(i);
 				Component* comp = reinterpret_cast<Component*>(&data->at(ent->entityComponents[i]));
-				
+
 				comp->ImguiWindowComponent();
 
 			}
@@ -32,7 +32,7 @@ void Entity::OnInspector(uint32_t entityID , Scene* scene)
 
 		if (ImGui::CollapsingHeader("Add Component"))
 		{
-			for (size_t i = 0; i < Component::GetComponentTypeInfos()->size(); i++)
+			for (uint32_t i = 0; i < Component::GetComponentTypeInfos()->size(); i++)
 			{
 
 				if (scene->HasComponent(i, ent))
@@ -56,7 +56,7 @@ void Entity::OnInspector(uint32_t entityID , Scene* scene)
 void Entity::OnSerialize(Scene* scene) const
 {
 
-	for (size_t i = 0; i < Component::GetNbrOfComponent(); i++)
+	for (uint32_t i = 0; i < Component::GetNbrOfComponent(); i++)
 	{
 		if (!scene->HasComponent(i, this))
 			continue;
@@ -66,7 +66,7 @@ void Entity::OnSerialize(Scene* scene) const
 
 
 
-		for (size_t k = 0; k < data->size() / Component::GetSizeOfComponent(i); k++)
+		for (uint32_t k = 0; k < data->size() / Component::GetSizeOfComponent(i); k++)
 		{
 			Component* ptr = reinterpret_cast<Component*> (&data->at(k));
 
@@ -82,7 +82,7 @@ void Entity::OnSerialize(Scene* scene) const
 
 void Entity::OnDeSerialize(Scene* scene) 
 {
-	for (size_t i = 0; i < Component::GetNbrOfComponent(); i++)
+	for (uint32_t i = 0; i < Component::GetNbrOfComponent(); i++)
 	{
 		if (!scene->HasComponent(i, this))
 			continue;
@@ -90,9 +90,9 @@ void Entity::OnDeSerialize(Scene* scene)
 
 		std::vector<uint8_t>* data = scene->GetComponentDataArray(i);
 
-		
 
-		for (size_t k = 0; k < data->size() / Component::GetSizeOfComponent(i); k++)
+
+		for (uint32_t k = 0; k < data->size() / Component::GetSizeOfComponent(i); k++)
 		{
 			Component* ptr = reinterpret_cast<Component*> (&data->at(k));
 
@@ -110,7 +110,7 @@ Entity::Entity()
 {
 	entityComponents.resize(Component::GetNbrOfComponent());
 
-	for (size_t i = 0; i < entityComponents.size(); i++)
+	for (uint32_t i = 0; i < entityComponents.size(); i++)
 	{
 		entityComponents[i] = ComponentNULL;
 	}
